add folders for middleware, models and routes. also add authentication and authorization
